+++ packages/sdks/output/react-native/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Builder.io React Native SDK Changelog (@builder.io/sdk-react-native)
 
+## 5.1.1
+
+### Patch Changes
+
+- fffde77: fix trackConversion method for sdks
+
 ## 5.1.0
 
 ### Minor Changes